start writing emitters
