+++ src/op_table.inl
@@ -1,4 +1,4 @@
-const OpInfo OPCODE_TABLE[OPCODE_SIZE] =
+constexpr OpInfo OPCODE_TABLE[OPCODE_SIZE] =
 {
     {op_group::reg_t,"mov %r, %r    ",2,{arg_type::dst_reg,arg_type::src_reg,arg_type::none}},
     {op_group::reg_t,"add %r, %r, %r",3,{arg_type::dst_reg,arg_type::src_reg,arg_type::src_reg}},
