+++ src/reg.cpp
@@ -1,0 +1,88 @@
+#include <ir.h>
+
+static constexpr u32 REG_FREE = SPECIAL_PURPOSE_REG_START - 1;
+static constexpr u32 TMP_END = REG_FREE - 1;
+static constexpr u32 REG_TMP_START = 0x00000000;
+
+void destroy_reg(Reg& ir_reg)
+{
+    destroy_arr(ir_reg.usage);
+}
+
+b32 is_sym(SymSlot s)
+{
+    return s.handle >= SYMBOL_START;
+}
+
+u32 tmp(u32 ir_reg)
+{
+    return ir_reg + REG_TMP_START;
+}
+
+// dont correct special regs
+b32 is_reg(SymSlot r)
+{
+    return r.handle < MACHINE_REG_SIZE;
+}
+
+b32 is_special_reg(SymSlot r)
+{
+    return r.handle >= SPECIAL_PURPOSE_REG_START && r.handle <= SPECIAL_PURPOSE_REG_START + SPECIAL_REG_SIZE;
+}
+
+b32 is_tmp(SymSlot s)
+{
+    return s.handle < TMP_END;
+}
+
+
+u32 slot_to_idx(SymSlot slot)
+{
+    return is_sym(slot)? sym_to_idx(slot) : slot.handle;
+}
+
+u32 gpr_count(u32 size)
+{
+    return size / GPR_SIZE;
+}
+
+
+void assign_reg_size(Reg& reg, u32 size)
+{
+    if(size > GPR_SIZE)
+    {
+        reg.size = GPR_SIZE;
+        reg.count = gpr_count(size);
+    }
+
+    else
+    {
+        reg.count = 1; 
+        reg.size = size;
+    }    
+}
+
+Reg make_reg(reg_kind kind,u32 size, u32 slot, b32 is_signed)
+{
+    Reg reg;
+    reg.kind = kind;
+
+    assign_reg_size(reg,size);
+
+    if(is_signed)
+    {
+        reg.flags |= SIGNED_FLAG;
+    }
+
+    reg.slot = {slot};
+
+    return reg;
+}
+
+
+void print(const Reg& reg)
+{
+    printf("offset: %x\n",reg.offset);
+    printf("location: %x\n\n",reg.location);    
+    printf("slot: %x\n",reg.slot.handle);
+}
